requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
python-dotenv
pydantic[email]
psycopg[binary]
PyJWT
pwdlib[argon2]
passlib
